requirements.txt
python-dotenv==1.0.1
replicate==1.0.4
Requests==2.32.3
slack_sdk==3.34.0
